services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
httpx==0.25.2
numpy==1.26.2
scikit-learn==1.3.2
google-generativeai==0.8.0
python-multipart==0.0.6
slowapi==0.1.9
